safeDeleteRecursively: Only delete non-writable files with force
This resembles the "rm" / "rm -f" command's behavior on Linux. However,
note that this is a change in default behavior compared to the previous
code. While "force" should usually not be required, use it at least in
all tests to be on the safe side. E.g. the Downloader's BableTest indeed
otherwise fails on Windows as a ready-only file inside a .git directory
cannot be deleted.

+++ utils/src/main/kotlin/Extensions.kt
@@ -60,7 +60,7 @@ fun File.realFile(): File = toPath().toRealPath().toFile()
  *
  * @throws IOException if the directory could not be deleted.
  */
-fun File.safeDeleteRecursively() {
+fun File.safeDeleteRecursively(force: Boolean = false) {
     if (!exists()) {
         return
     }
@@ -82,6 +82,8 @@ fun File.safeDeleteRecursively() {
             try {
                 Files.delete(file)
             } catch (e: java.nio.file.AccessDeniedException) {
+                if (!force) throw e
+
                 if (file.toFile().setWritable(true)) {
                     // Try again.
                     Files.delete(file)
